pyproject.toml
[project]
name = "tapesmith-option-a"
version = "0.1.0"
description = "TapeSmith MCP Server — Direct MCP tools for AI Playground"
requires-python = ">=3.10"
dependencies = [
    "mcp[cli]>=1.2.0",
    "databricks-sdk>=0.20.0",
    "databricks-langchain>=0.1.0",
    "langchain-core>=0.2.0",
    "pandas>=2.0.0",
    "numpy>=1.24.0",
    "openpyxl>=3.0.0",
    "requests>=2.28.0",
    "python-dateutil>=2.8.0",
]

[project.scripts]
tapesmith-mcp-a = "server.main:main"

[tool.setuptools.packages.find]
include = ["server*", "processing*"]
exclude = ["notebooks*"]

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

